requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
requests
python-dotenv
